Create logic in service layer Early changes in book borrow logic Add todo to security configuration - need edit permissions

+++ src/main/java/mTrepka/libary/service/BookServiceImpl.java
@@ -2,12 +2,12 @@ package mTrepka.libary.service;
 
 import lombok.RequiredArgsConstructor;
 import mTrepka.libary.domain.Book;
+import mTrepka.libary.domain.BorrowHistory;
 import mTrepka.libary.domain.User;
 import mTrepka.libary.repository.BookRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.sql.Connection;
+import java.time.LocalDate;
 import java.util.List;
 
 /**
@@ -17,6 +17,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BookServiceImpl implements BookService{
 	private final BookRepository bookRepository;
+    private final UserService userService;
 
     @Override
     public List<Book> getAllBooks() {
@@ -30,17 +31,33 @@ public class BookServiceImpl implements BookService{
 
     @Override
     public List<Book> findByOwnerid(User ownerid) {
-        return bookRepository.findByOwnerid(ownerid);
+        return bookRepository.findByOwner(ownerid);
     }
 
-    @Override
-    public void saveBook(Book book) {
-        bookRepository.save(book);
-    }
 
     @Override
     public List<Book> findAllFreeBook() {
         return bookRepository.findFreeBook();
+    }
+
+    @Override
+    public void addBook(Book book) {
+        bookRepository.save(book);
+    }
+
+    @Override
+    public void borrowBook(Long bookId) {
+        Book book = bookRepository.findById(bookId);
+        if (book.getOwner() == null) {
+            User user = userService.getCurrentUser();
+            BorrowHistory borrowHistory = new BorrowHistory();
+            borrowHistory.setBorrow_date(LocalDate.now());
+            borrowHistory.setBookBorrow(book);
+            borrowHistory.setUserBorrow(user);
+            book.setCurrentBorrow(borrowHistory);
+            book.setOwner(user);
+            bookRepository.save(book);
+        }
     }
 
     @Override
@@ -49,7 +66,30 @@ public class BookServiceImpl implements BookService{
     }
 
     @Override
-    public void removeBook(Book book) {
-        bookRepository.delete(book);
+    public String removeBook(Book book) {
+        if (book.getOwner() == null) {
+            bookRepository.delete(book);
+            return "Book has been removed";
+        } else {
+            return "You can't remove book becouse is borrowed";
+        }
+
     }
+
+    @Override
+    public Book editBookAndSave(Book bookData, long bookId) {
+        Book currentBook = bookRepository.findById(bookId);
+        editBookFromTemplate(currentBook, bookData);
+        return currentBook;
+    }
+
+    private void editBookFromTemplate(Book currentBook, Book bookData) {
+        currentBook.setAuthor(bookData.getAuthor());
+        currentBook.setName(bookData.getName());
+        currentBook.setIsbn(bookData.getIsbn());
+        currentBook.setDescription(bookData.getDescription());
+        currentBook.setPageAmount(bookData.getPageAmount());
+        currentBook.setReleaseDate(bookData.getReleaseDate());
+    }
+
 }
